refactoring, components moved to dirs
