+++ weather-fetcher-stats-ui/src/components/request-stats/modals/RequestsVisualization.tsx
@@ -1,10 +1,10 @@
-import {RequestData} from "./models/RequestData";
+import {RequestData} from "../../../types/RequestData";
 import RequestsPerHourLineChartModal from "./RequestsPerHourLineChartModal";
 import CodeRatioPieChartModal from "./CodeRatioPieChartModal";
 import React from "react";
-import {useStyles} from "./constants/styles";
+import {useStyles} from "../../../constants/styles";
 
-export const VisualizationBlock = ({requests, disabled}: { requests: RequestData[], disabled: boolean }) => (
+export const RequestsVisualization = ({requests, disabled}: { requests: RequestData[], disabled: boolean }) => (
     <div className={useStyles().visualizationBlock}>
         <span>Visualization:</span>
         <RequestsPerHourLineChartModal requests={requests} disabled={disabled}/>
