+++ weather-fetcher-stats-ui/src/components/request-stats/table/RequestTable.tsx
@@ -1,12 +1,12 @@
-import {Alert, CircularProgress, Paper, Table, TableCell} from "@mui/material";
+import {Alert, CircularProgress, Paper, Table, TableBody, TableCell, TableRow} from "@mui/material";
 import {RequestTableHead} from "./RequestTableHead";
 import {RequestTableItem} from "./RequestTableItem";
 import {RequestTablePagination} from "./RequestTablePagination";
 import React from "react";
-import {useStyles} from "./constants/styles";
-import {PaginationProps} from "./models/PaginationProps";
-import {RequestData} from "./models/RequestData";
-import {TableSortProps} from "./models/TableSortProps";
+import {useStyles} from "../../../constants/styles";
+import {PaginationProps} from "../../../types/props/PaginationProps";
+import {RequestData} from "../../../types/RequestData";
+import {TableSortProps} from "../../../types/props/TableSortProps";
 
 interface RequestTableProps {
     paginationProps: PaginationProps,
@@ -15,7 +15,6 @@ interface RequestTableProps {
 }
 
 export const RequestTable = ({paginationProps, tableSortProps, appData}: RequestTableProps) => {
-
     const classes = useStyles()
 
     const handleSort = (field: string) => {
@@ -23,8 +22,7 @@ export const RequestTable = ({paginationProps, tableSortProps, appData}: Request
         tableSortProps.setSortField(field);
         tableSortProps.setSortOrder(newSortOrder);
     };
-    console.log(appData.error)
-    console.log(appData.loading)
+
     return (
         <Paper elevation={5} className={classes.root}>
             {appData.error && (
@@ -38,15 +36,20 @@ export const RequestTable = ({paginationProps, tableSortProps, appData}: Request
                     sortOrder={tableSortProps.sortOrder}
                     handleSort={handleSort}
                 />
-                {appData.loading ? (
-                    <TableCell colSpan={7} align="center">
-                        <CircularProgress/>
-                    </TableCell>
-                ) : (
-                    appData?.requests.map((request, index) => (
-                        <RequestTableItem request={request} index={index} key={request.id}/>
-                    ))
-                )}
+                <TableBody>
+                    {appData.loading ? (
+                        <TableRow>
+                            <TableCell colSpan={7} align="center">
+                                <CircularProgress/>
+                            </TableCell>
+                        </TableRow>
+                    ) : (
+                        appData?.requests.map((request, index) => (
+                            <RequestTableItem request={request} index={index} key={request.id}/>
+                        ))
+                    )}
+                </TableBody>
+
             </Table>
             <RequestTablePagination
                 page={paginationProps.page}
