+++ weather-fetcher-stats-ui/src/hooks/useFetchRequestStats.ts
@@ -1,16 +1,8 @@
 import {useCallback, useEffect, useState} from "react";
-import {config} from "../constants/config";
-import {ApiResponse} from "../models/RequestApiResponse";
-import {RequestData} from "../models/RequestData";
+import {ApiResponse} from "../types/RequestApiResponse";
+import {RequestData} from "../types/RequestData";
 
-interface RequestStatsProps {
-    page: number,
-    rowsPerPage: number,
-    sortField: string,
-    sortOrder: 'asc' | 'desc'
-}
-
-export const useRequestStats = (requestStatsProps: RequestStatsProps) => {
+export const useFetchRequestStats = (url: string) => {
 
     const [appData, setAppData] =
         useState<{ requests: RequestData[], loading: boolean, error: string }>({
@@ -24,12 +16,8 @@ export const useRequestStats = (requestStatsProps: RequestStatsProps) => {
     const handleFetchRequestsCallback = useCallback(() => {
         const fetchRequests = async () => {
             setAppData({...appData, loading: true})
-            const response =
-                await fetch(
-                    `${config.WF_STATS_URL}:${config.WF_STATS_PORT}${config.LIST_REQUEST_STATS_ENDPOINT}?page=${requestStatsProps.page}&size=${requestStatsProps.rowsPerPage}&sortBy=${requestStatsProps.sortField}&order=${requestStatsProps.sortOrder}`
-                );
+            const response = await fetch(url);
             const data: ApiResponse = await response.json();
-
             setTotalElements(data.totalElements);
             setAppData({...appData, requests: data.content, loading: false})
         };
@@ -38,7 +26,7 @@ export const useRequestStats = (requestStatsProps: RequestStatsProps) => {
             setAppData({...appData, error: `${reason.message}. Try later`})
         })
 
-    }, [requestStatsProps.page, requestStatsProps.rowsPerPage, requestStatsProps.sortField, requestStatsProps.sortOrder]);
+    }, [url]);
 
     useEffect(() => {
         handleFetchRequestsCallback()
