+++ weather-fetcher-stats-ui/src/index.tsx
@@ -2,15 +2,12 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 import {BrowserRouter} from "react-router-dom";
-import {ModalState} from './context/ModalContext';
 
 ReactDOM.render(
     <BrowserRouter>
-        <ModalState>
-            <React.StrictMode>
-                <App/>
-            </React.StrictMode>
-        </ModalState>
+        <React.StrictMode>
+            <App/>
+        </React.StrictMode>
     </BrowserRouter>,
     document.getElementById('root') as HTMLElement
 );
