+++ weather-fetcher-stats-ui/src/types/RequestApiResponse.ts
