+++ weather-fetcher-stats-ui/src/types/RequestData.ts
