+++ weather-fetcher-stats-ui/src/types/props/PaginationProps.ts
