+++ weather-fetcher-stats-ui/src/types/props/TableSortProps.ts
